+++ tests/e2e/matrix/internal/transport/bindings_service/echo_header_metadata.go
@@ -31,8 +31,6 @@ func EchoHeaderMetadata(svcCtx *svccontext.ServiceContext) http.HandlerFunc {
 			return
 		}
 		w.Header().Set("ETag", resp.Etag)
-		w.Header().Set("Content-Type", "application/json; charset=utf-8")
-		w.WriteHeader(http.StatusCreated)
-		_ = server.JSON().Encode(w, resp)
+		server.WriteResponse(w, r, http.StatusCreated, resp)
 	}
 }